experiments/real_world/scrna_commitment_tier2p11/requirements.txt
numpy>=1.26
pandas>=2.0
scipy>=1.11
pyyaml>=6.0
matplotlib>=3.8
pyarrow>=14.0
scanpy>=1.9.6
anndata>=0.10
scikit-learn>=1.4
leidenalg>=0.11
igraph>=0.11
openpyxl>=3.1
